swagger dcoumentatio update

+++ StarWars.Api/Controllers/EpisodesController.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using StarWars.Common.Model;
 using StarWars.Domain.Repositories;
 
@@ -27,22 +28,45 @@ namespace StarWars.Api.Controllers
             _mapper = mapper;
         }
 
+        /// <summary>
+        /// List of all episodes
+        /// </summary>
+        /// <returns></returns>
         [HttpGet]
         public async Task<IEnumerable<Episode>> Get()
         {
             return await _episodeRepository.GetEpisodeAsync();
         }
 
+        /// <summary>
+        /// Delete episode with selected id
+        /// </summary>
+        /// <param name="episodeId">Episode id to delete</param>
+        /// <returns>ActionResult</returns>
         [HttpDelete("{episodeId}")]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<IActionResult> Delete(int episodeId)
         {
+            if (!await _episodeRepository.EpisodeExistsAsync(episodeId))
+            {
+                return NotFound();
+            }
             _episodeRepository.DeleteEpisode(episodeId);
             await _episodeRepository.SaveChangesAsync();
             return NoContent();
 
         }
 
+        /// <summary>
+        /// Update episode
+        /// </summary>
+        /// <param name="episodeId">Episode id to update</param>
+        /// <param name="episodeDto">new episode data</param>
+        /// <returns></returns>
         [HttpPatch("{episodeId}")]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status201Created)]
         public async Task<IActionResult> Update(int episodeId, [FromBody] EpisodeDto episodeDto)
         {
             if (await _episodeRepository.EpisodeNameExistsAsync(episodeDto.EpisodeName))
@@ -60,6 +84,12 @@ namespace StarWars.Api.Controllers
 
         }
 
+        /// <summary>
+        /// Get episode by Id
+        /// </summary>
+        /// <param name="episodeId">selected episode id</param>
+        /// <returns>Episode</returns>
+
         [HttpGet("{episodeId}")]
         public async Task<ActionResult<Episode>> GetEpisode(
 
@@ -68,6 +98,13 @@ namespace StarWars.Api.Controllers
             return await _episodeRepository.GetEpisodeAsync(episodeId);
         }
 
+
+        /// <summary>
+        /// Create new episode
+        /// </summary>
+        /// <param name="episode">Episode name</param>
+        /// <returns>ActionResult</returns>
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPost()]
         public async Task<ActionResult<Episode>> CreateEpisode([FromBody] EpisodeDto episode)
         {
